Fixed some refactor bugs: now working with poetry run
